+++ example/main.dart
@@ -3,14 +3,15 @@ import 'package:routingkit/routingkit.dart';
 void main() {
   final router = createRouter<String>();
 
-  addRoute(router, 'GET', '/path', 'static route');
-  addRoute(router, 'POST', '/path/:name', 'name route');
-  addRoute(router, 'GET', '/path/foo/**', 'wildcard route');
-  addRoute(router, 'GET', '/path/foo/**:name', 'named wildcard route');
+  router.add('GET', '/path', 'static route');
+  router.add('POST', '/path/:name', 'name route');
+  router.add('GET', '/path/foo/**', 'wildcard route');
+  router.add('GET', '/path/foo/**:name', 'named wildcard route');
 
-  print(findRoute(router, 'GET', '/path')); // => { data: static route }
-  print(findRoute(router, 'POST', '/path/cady')); // => { data: name route, }
-  print(findRoute(
-      router, 'GET', '/path/foo/bar/baz')); // => { data: wildcard route }
-  print(findRoute(router, 'GET', '/')); // => null, not found.
+  print(router.find('GET', '/path')); // => { data: static route }
+  print(router.find(
+      'POST', '/path/cady')); // => { data: name route, params: {name: cady} }
+  print(router.find('GET',
+      '/path/foo/bar/baz')); // => { data: wildcard route, params: {_: bar/baz} }
+  print(router.find('GET', '/')); // => null, not found.
 }